requirements.txt
langchain
langchain_community 
langchain-openai
qdrant-client 
docker 
python-dotenv
gradio
paramiko # for ssh remote
fastembed
pypdf2